+++ packages/leancode_cubit_utils/CHANGELOG.md
@@ -4,7 +4,7 @@
 * **BREAKING CHANGE**: Remove `TData` generic parameter from `RequestCubit`.
 * Add `map` method to `RequestState`.
 * Add `refreshing` callback to `RequestCubitBuilder`.
-* **BREAKING CHANGE**: Make `data` in `RequestRefreshState` non-nullable.
+* **BREAKING CHANGE**: Make `data` in `RequestRefreshingState` non-nullable.
 * **BREAKING CHANGE**: Add `data` to `RequestEmptyState`.
 
 ## 0.4.2